feat: web specific layout
