+++ src/types/WebLayoutLink.ts
@@ -1,0 +1,5 @@
+export type WebLayoutLink = {
+  href: string;
+  name?: string;
+  isActive: (path: string) => boolean;
+};
